+++ appstore-core/domain/src/main/java/com/example/delivery/model/dto/response/DeliverySummaryResponse.java
@@ -8,7 +8,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -29,9 +28,9 @@ public class DeliverySummaryResponse {
 
     private String deliveryOptions;
 
-    private BigDecimal deliveryFee; //BigDemical
+    private double deliveryFee; //BigDemical
 
-    private DeliverySummaryResponse(String reservationNumber, LocalDateTime deliveryDate, DeliveryAddress deliveryAddress, Vehicle vehicle, Integer stopOverCount, String deliveryOptions, BigDecimal deliveryFee) {
+    private DeliverySummaryResponse(String reservationNumber, LocalDateTime deliveryDate, DeliveryAddress deliveryAddress, Vehicle vehicle, Integer stopOverCount, String deliveryOptions, double deliveryFee) {
         this.reservationNumber = reservationNumber;
         this.deliveryDate = deliveryDate;
         this.deliveryAddress = deliveryAddress;
